--- src/main/java/nl/uva/deepstream/App.java
@@ -1,48 +1,0 @@
-package nl.uva.deepstream;
-
-import org.nd4j.linalg.api.ndarray.INDArray;
-import org.nd4j.linalg.factory.NDArrayFactory;
-import org.nd4j.linalg.factory.Nd4j;
-import org.nd4j.linalg.indexing.INDArrayIndex;
-
-
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-//    INDArray arr;
-
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-
-        short a = 25;
-        short b = (short)-a;
-        System.out.println(a + ", " + b);
-
-//        new NDArrayFactory().create();
-//
-//        INDArray nd = Nd4j.create(new int[]{1, 2, 3, 4}, new int[]{2, 2});
-//        System.out.println(nd);
-//
-//        System.out.println(nd.)
-//
-//        nd.getColumn(1).subi(5);
-//
-////        nd.put(0, 1, 5);
-//
-//        System.out.println(nd);
-////
-//        nd.assign(0);
-//
-//        System.out.println(nd);
-//
-//        nd.getRow(1).assign(6);
-//        System.out.println(nd);
-
-
-    }
-}
