+++ src/main/java/nl/uva/deepstream/event_modules/BaseSpikeEvent.java
@@ -1,0 +1,13 @@
+package nl.uva.deepstream.event_modules;
+
+/**
+ * Created by peter on 2/24/16.
+ */
+public class BaseSpikeEvent{
+
+    public final int src;
+
+    public BaseSpikeEvent(int src) {
+        this.src = src;
+    }
+}
