+++ src/main/java/nl/uva/deepstream/event_modules/NonResettingRectVarQueueHerder.java
@@ -1,0 +1,16 @@
+package nl.uva.deepstream.event_modules;
+
+/**
+ * Created by peter on 2/24/16.
+ */
+public class NonResettingRectVarQueueHerder extends RectVarQueueHerder {
+    /* A version of the varqueueherder where phi never resets */
+
+    public NonResettingRectVarQueueHerder(int n_units) {
+        super(n_units);
+    }
+
+    @Override
+    public void reset(){
+    }
+}
