+++ src/main/java/nl/uva/deepstream/event_modules/StatelessBaseEventModule.java
@@ -1,0 +1,13 @@
+package nl.uva.deepstream.event_modules;
+
+/**
+ * Created by peter on 2/24/16.
+ */
+abstract class StatelessBaseEventModule<EventType> extends BaseEventModule<EventType> {
+    /* Started this because of bugs caused by forgetting to reset.  Now you have to explicitly not reset */
+
+    @Override
+    public void reset(){
+    }
+
+}
