+++ src/main/java/nl/uva/deepstream/interfaces/Resettable.java
@@ -1,0 +1,9 @@
+package nl.uva.deepstream.interfaces;
+
+/**
+ * Created by peter on 2/24/16.
+ */
+public interface Resettable{
+
+    void reset();
+}
