+++ src/main/java/nl/uva/deepstream/interfaces/Router.java
@@ -1,0 +1,14 @@
+package nl.uva.deepstream.interfaces;
+
+import java.util.function.Consumer;
+
+/**
+ * Created by peter on 2/24/16.
+ */
+public interface Router extends EventHandler, Resettable{
+
+    <EventType> void add_binding(EventModule<EventType> src_module, Consumer<EventType> method_reference);
+
+    <EventType> void handle_event(EventType ev, Consumer<EventType> consumer);
+
+}
